requirements.txt
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
